CadastroUsuario
003

+++ Financas/src/br/com/fiap/financas/dao/DataSource.java
@@ -12,7 +12,7 @@ public class DataSource extends SQLiteOpenHelper {
 	protected static final String TABLE_MOVIMENTACAO = "tb_movimentacao";
 	protected static final String TABLE_CATEGORIA = "tb_categoria";
 	protected static final String TABLE_ORIGEM = "tb_origem";
-	private static final int DATABASE_VERSION = 1;
+	private static final int DATABASE_VERSION = 2;
 
 	Context context;
 	SQLiteDatabase db;
@@ -20,6 +20,10 @@ public class DataSource extends SQLiteOpenHelper {
 	public DataSource(Context context) {
 
 		super(context, DATABASE_FINANCAS, null, DATABASE_VERSION);
+		this.context  = context ;
+		this.db = getWritableDatabase();
+		//onUpgrade(db, 2, DATABASE_VERSION);
+		onUpgrade(db, 2, 1);
 	}
 
 	@Override
@@ -27,7 +31,7 @@ public class DataSource extends SQLiteOpenHelper {
 
 		db.execSQL("CREATE TABLE "
 				+ TABLE_USUARIO
-				+ " (id_usuario INTEGER PRIMARY KEY AUTOINCREMENT, nome TEXT, usuario TEXT, senha TEXT, flag_ativo VARCHAR)");
+				+ " (id_usuario INTEGER PRIMARY KEY AUTOINCREMENT, nome TEXT, usuario TEXT, senha TEXT, email TEXT, telefone REAL, flag_ativo VARCHAR)");
 
 		db.execSQL("CREATE TABLE "
 				+ TABLE_MOVIMENTACAO
@@ -47,11 +51,16 @@ public class DataSource extends SQLiteOpenHelper {
 	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 		Log.w("Finanças",
 				"Upgrading database, this will drop tables and recreate.");
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_MOVIMENTACAO);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_USUARIO);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_CATEGORIA);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_ORIGEM);
-		onCreate(db);
+		if(newVersion > oldVersion){
+			db.execSQL("DROP TABLE IF EXISTS " + TABLE_MOVIMENTACAO);
+			db.execSQL("DROP TABLE IF EXISTS " + TABLE_USUARIO);
+			db.execSQL("DROP TABLE IF EXISTS " + TABLE_CATEGORIA);
+			db.execSQL("DROP TABLE IF EXISTS " + TABLE_ORIGEM);
+			onCreate(db);
+		}
 	}
 
+	public void close(){
+		this.db.close();
+	}
 }
